Add game 1 challenge

+++ packages/contracts/src/codegen/world/ITowerSystem.sol
@@ -11,9 +11,9 @@ pragma solidity >=0.8.24;
 interface ITowerSystem {
   function app__getTowerSystemAddress() external view returns (address);
 
-  function app__installTower(bytes32 potentialGameId, bool projectile, int8 x, int8 y) external returns (bytes32);
+  function app__installTower(bytes32 potentialGameId, bool projectile, int16 x, int16 y) external returns (bytes32);
 
-  function app__moveTower(bytes32 potentialGameId, bytes32 towerId, int8 x, int8 y) external returns (bytes32);
+  function app__moveTower(bytes32 potentialGameId, bytes32 towerId, int16 x, int16 y) external returns (bytes32);
 
   function app__modifyTowerSystem(
     bytes32 towerId,
