+++ packages/contracts/src/defaultLogicContracts/DefaultProjectileLogicLeft.sol
@@ -2,7 +2,7 @@
 pragma solidity >=0.8.24;
 
 contract DefaultProjectileLogicLeft {
-  function getNextProjectilePosition(int8 x, int8 y) public pure returns (int8, int8) {
+  function getNextProjectilePosition(int16 x, int16 y) public pure returns (int16, int16) {
     return (x + 1, y);
   }
 }
